Move isV1 check inside of catch scope
In the case `web3` is not defined in the browser, the value of `window.web3` is `null` which causes an exception to be raised at a browser level. By moving it inside the try/catch scope we can swallow the exception and return the correct "your browser does not support web3" screen.